Main Verilog Files
These are the two main files used to program a De1-SoC to take input from a PS/2 keyboard and interpret it as a piano. It outputs sound through the 3.5mm audio jack on the board. These files call modules from audio and PS/2 interface files that I will not be including.